\name{BivGEV.fit}
\alias{BivGEV.fit}

\title{
Recall the optimization function
}
\description{
Internal function that recall the optimization function. \code{BivGEV.fit} function is a wrapper from \code{SemiParBIV.fit()} in \code{GJRM}.
}

\author{
Raffaella Calabrese - \email{raffaella.calabrese@ed.ac.uk};
Silvia Angela Osmetti - \email{silvia.osmetti@unicatt.it}; 
Luca Zanin - \email{luca.zanin@studio.unibo.it};
}

\seealso{
\code{\link{BivGEV}}
}
\keyword{Bivariate Generalised Extreme Value Model}