\name{BivGEVOptim}
\alias{BivGEVOptim}
\title{
This is a function of optimization when estimating joint imbalanced binary margins (internal function)
}
\description{
Internal function for optimization when estimating joint imbalanced binary margins. %\code{BivGEVOptim} function is a wrapper from \code{bprobgHs()} in \code{GJRM}.
}

\author{
Raffaella Calabrese - \email{raffaella.calabrese@ed.ac.uk};
Silvia Angela Osmetti - \email{silvia.osmetti@unicatt.it}; 
Mantainer: Luca Zanin - \email{luca.zanin@studio.unibo.it};
}

\seealso{
\code{\link{BivGEV}}
}


\keyword{Bivariate Generalised Extreme Value Model}