\name{BivGEVOptimSS}
\alias{BivGEVOptimSS}

\title{
Optimization when estimating joint imbalanced binary margins in the presence of sample selection
}
\description{
Internal function for optimization when estimating joint imbalanced binary margins in the presence of sample selection. \code{BivGEVOptimSS} function is a wrapper from \code{bprobgHsSS()} in \code{GJRM}.
}

\author{
Raffaella Calabrese - \email{raffaella.calabrese@ed.ac.uk};
Silvia Angela Osmetti - \email{silvia.osmetti@unicatt.it}; 
Luca Zanin - \email{luca.zanin@studio.unibo.it};
}

\seealso{
\code{\link{BivGEV}}
}

\keyword{Bivariate Generalised Extreme Value Model}