\name{print.summary.BivGEV}
\alias{print.summary.BivGEV}

\title{
Print summary statistics of the \code{BivGEV} model 
}
\description{
Print summary statistics of the \code{BivGEV} model 
}
\usage{
\method{print.summary}{BivGEV}(x, 
digits = max(3, getOption("digits") - 3), 
signif.stars = getOption("show.signif.stars"), 
...)
}

\arguments{
  \item{x}{
A fitted \code{BivGEV} object.
}
  \item{digits}{
Number of digits printed in output Tables.
}
  \item{signif.stars}{
The significance stars are printed by default.
}
\item{...}{Other arguments.}
}



\author{
Raffaella Calabrese - \email{raffaella.calabrese@ed.ac.uk};
Silvia Angela Osmetti - \email{silvia.osmetti@unicatt.it}; 
Mantainer: Luca Zanin - \email{luca.zanin@studio.unibo.it};
}

\seealso{
\code{\link{BivGEV}}
}


\keyword{Bivariate Generalised Extreme Value Model}
