\name{train.test}
\alias{train.test}

\title{
The function split the sample in training and test set.
}
\description{
The function split the sample in training and test set.
}
\usage{
train.test(data, perc = 0.75, seed = 100)
}

\arguments{
  \item{data}{
The data frame that includes the variables of the model.
}
  \item{perc}{
Setting the percentage of units in the training sample. 0.75 is the default value.
}
  \item{seed}{
Setting the seed. 100 is the default value.
}
}

\author{
Raffaella Calabrese - \email{raffaella.calabrese@ed.ac.uk};
Silvia Angela Osmetti - \email{silvia.osmetti@unicatt.it}; 
Maintainer: Luca Zanin - \email{luca.zanin@studio.unibo.it};
}

\seealso{
\code{\link{BivGEV}}
}

\keyword{Bivariate Generalised Extreme Value Model}


