\name{work.fit}
\alias{work.fit}

\title{
Internal Function
}
\description{
Internal Function
}
%\usage{
%work.fit(fit)
%}

\author{
Raffaella Calabrese - \email{raffaella.calabrese@ed.ac.uk};
Silvia Angela Osmetti - \email{silvia.osmetti@unicatt.it}; 
Luca Zanin - \email{luca.zanin@studio.unibo.it};
}

\keyword{Bivariate Generalised Extreme Value Model}