\name{BivGEVFit.post}
\alias{BivGEVFit.post}

\title{
This is a function for computing post estimation quantities (internal function)
}
\description{
Internal function for computing post estimation quantities. %\code{BivGEVFit.post} function is a wrapper from \code{SemiParBIV.fit.post()} in \code{GJRM}.
}

\author{
Raffaella Calabrese - \email{raffaella.calabrese@ed.ac.uk};
Silvia Angela Osmetti - \email{silvia.osmetti@unicatt.it}; 
Mantainer: Luca Zanin - \email{luca.zanin@studio.unibo.it};
}

\seealso{
\code{\link{BivGEV}}
}

\keyword{Bivariate Generalised Extreme Value Model}