\name{diagnostics}
\alias{diagnostics}

\title{
Diagnostics on model's convergence
}
\description{
Diagnostics on model's convergence
}
\usage{
diagnostics(x)
}

\arguments{
  \item{x}{
A fitted \code{BivGEV} object.
}
}


\author{
Raffaella Calabrese - \email{raffaella.calabrese@ed.ac.uk};
Silvia Angela Osmetti - \email{silvia.osmetti@unicatt.it}; 
Luca Zanin - \email{luca.zanin@studio.unibo.it};
}

\seealso{
\code{\link{BivGEV}}
}


\keyword{Bivariate Generalised Extreme Value Model}
