\name{logLik.BivGEV}
\alias{logLik.BivGEV}

\title{
Function extract the log-likelihood for a BivGEV model
}
\description{
The function extract the log-likelihood for a \code{BivGEV} model.  
}
\usage{
\method{logLik}{BivGEV}(object, ...)
}

\arguments{
  \item{object}{
A fitted \code{BivGEV} object.
}
\item{...}{Other arguments.}
}

\author{
Raffaella Calabrese - \email{raffaella.calabrese@ed.ac.uk};
Silvia Angela Osmetti - \email{silvia.osmetti@unicatt.it}; 
Luca Zanin - \email{luca.zanin@studio.unibo.it};
}

\seealso{
\code{\link{BivGEV}}
}


\keyword{Bivariate Generalised Extreme Value Model}